:lipstick: Make buttons `button` elements for a11y
Button elements are natively tabbable and supports enter and space bar clicks. I have chosen to set the `tabindex` of the second button (the `code` wrapper) to `-1`, as this will remove it from the tab tree – we don't need more than one button per Gitmoji when tabbing through. This might be subject to change, as we might want to set the `code` wrapper as the main button.

This PR also resets any button styles in order to make the old and the new versions look exactly the same.