:bug: Render unicode gitmojis as image by default (#753)
Some emojis are rendered as text by default, other as image. The result
is that some gitmojis are different / a bit ugly depending on the font
used. This concerns only the unicode version of the gitmoji, not the
:code: one.

Emojis are rendered either by text or picture. Variation selectors
VS15 and VS16 define if text or picture representation should be used.

According to [0], the following gitmojis are affected by those selectors
and may be rendered as text:

- alembic
- alien
- ambulance
- *arrow-down
- *arrow-up
- building-construction
- card-file-box
- *coffin
- *label
- lock
- mag
- package
- *pencil2
- *recycle
- rewind
- wastebasket
- *wheelchair
- *zap

Out of those, 8 already have VS16 (\ufe0f), and will be rendered as
image. This commit unifies the 10 other ones by adding the selector.

[0]: https://www.unicode.org/Public/13.0.0/ucd/emoji/emoji-variation-sequences.txt